requirements.txt
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
bidict==0.23.1
billiard==4.2.0
blinker==1.8.2
blis==0.7.11
cachelib==0.13.0
catalogue==2.0.10
celery==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpathlib==0.16.0
confection==0.1.5
configobj==5.0.8
configparser==7.0.0
contourpy==1.1.1
cycler==0.12.1
cymem==2.0.8
distro==1.9.0
dnspython==2.6.1
en-core-web-sm==3.7.1
etelemetry==0.3.1
eventlet==0.36.1
exceptiongroup==1.2.1
filelock==3.14.0
fitz==0.0.1.dev2
flask==3.0.3
Flask-Cors==4.0.1
flask-session==0.8.0
Flask-SocketIO==5.3.6
fonttools==4.53.0
fsspec==2024.5.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.23.2
idna==3.7
importlib-metadata==7.1.0
importlib-resources==6.4.0
isodate==0.6.1
itsdangerous==2.2.0
jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
kombu==5.3.7
langcodes==3.4.0
language-data==1.2.0
llvmlite==0.41.1
looseversion==1.3.0
lxml==5.2.2
marisa-trie==1.1.1
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
msgspec==0.18.6
murmurhash==1.0.10
networkx==3.1
nibabel==5.2.1
nipype==1.8.6
numba==0.58.1
numpy==1.24.4
openai==1.30.5
outcome==1.3.0.post0
packaging==24.0
pandas==2.0.3
pathlib==1.0.1
pillow==10.3.0
pinecone==4.0.0
pinecone-client==4.1.0
preshed==3.0.9
prompt-toolkit==3.0.45
prov==2.0.0
pydantic==2.7.2
pydantic-core==2.18.3
pydot==2.0.0
PyMuPDF==1.24.5
PyMuPDFb==1.24.3
pynndescent==0.5.12
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-engineio==4.9.1
python-socketio==5.11.2
pytz==2024.1
pyxnat==1.6.2
PyYAML==6.0.1
rdflib==6.3.2
redis==5.0.4
regex==2024.5.15
replicate==0.26.0
requests==2.32.3
safetensors==0.4.3
scikit-learn==1.3.2
scipy==1.10.1
selenium==4.21.0
sentence-transformers==3.0.0
simple-websocket==1.0.0
simplejson==3.19.2
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.12.1
thinc==8.2.3
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.2.2
tqdm==4.66.4
traits==6.3.2
transformers==4.41.2
trio==0.25.1
trio-websocket==0.11.1
typer==0.9.4
typing-extensions==4.12.0
tzdata==2024.1
umap-learn==0.5.6
urllib3==2.2.1
vine==5.1.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.3.4
werkzeug==3.0.3
wsproto==1.2.0
zipp==3.19.0
